Removing AuthenticatorFactory interface
