+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/AuthenticatorFactoryTest.java
@@ -1,4 +1,4 @@
-package com.symphony.bdk.core.auth.impl;
+package com.symphony.bdk.core.auth;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -6,11 +6,13 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.symphony.bdk.core.auth.BotAuthenticator;
-import com.symphony.bdk.core.auth.ExtensionAppAuthenticator;
-import com.symphony.bdk.core.auth.AuthenticatorFactory;
-import com.symphony.bdk.core.auth.OboAuthenticator;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
+import com.symphony.bdk.core.auth.impl.BotAuthenticatorCertImpl;
+import com.symphony.bdk.core.auth.impl.BotAuthenticatorRsaImpl;
+import com.symphony.bdk.core.auth.impl.ExtensionAppAuthenticatorCertImpl;
+import com.symphony.bdk.core.auth.impl.ExtensionAppAuthenticatorRsaImpl;
+import com.symphony.bdk.core.auth.impl.OboAuthenticatorCertImpl;
+import com.symphony.bdk.core.auth.impl.OboAuthenticatorRsaImpl;
 import com.symphony.bdk.core.client.ApiClientFactory;
 import com.symphony.bdk.core.config.model.BdkCertificateConfig;
 import com.symphony.bdk.core.config.model.BdkConfig;
@@ -31,9 +33,9 @@ import java.util.UUID;
 import java.util.function.Supplier;
 
 /**
- * Test class for the {@link AuthenticatorFactoryImpl}.
+ * Test class for the {@link AuthenticatorFactory}.
  */
-class AuthenticatorFactoryImplTest {
+class AuthenticatorFactoryTest {
 
   private static final ApiClient DUMMY_API_CLIENT = new ApiClientBuilderJersey2().build();
   private static final String RSA_PRIVATE_KEY = RsaTestHelper.generatePrivateKeyAsString();
@@ -58,7 +60,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -68,7 +70,7 @@ class AuthenticatorFactoryImplTest {
   void testGetBotAuthenticatorWithValidPrivateKeyInClasspath() throws AuthInitializationException {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/private-key.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -78,7 +80,7 @@ class AuthenticatorFactoryImplTest {
   void testGetBotAuthenticatorWithPrivateKeyNotFoundInClasspath() {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/notfound.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, () -> factory.getBotAuthenticator());
   }
 
@@ -88,7 +90,7 @@ class AuthenticatorFactoryImplTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent(RSA_PRIVATE_KEY.getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
 
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -100,7 +102,7 @@ class AuthenticatorFactoryImplTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent("invalid-key".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -114,7 +116,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -127,7 +129,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -136,7 +138,7 @@ class AuthenticatorFactoryImplTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -149,7 +151,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final OboAuthenticator oboAuth = factory.getOboAuthenticator();
     assertNotNull(oboAuth);
     assertEquals(OboAuthenticatorRsaImpl.class, oboAuth.getClass());
@@ -161,7 +163,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     BotAuthenticator botAuthenticator = factory.getBotAuthenticator();
     assertNotNull(botAuthenticator);
     assertEquals(BotAuthenticatorCertImpl.class, botAuthenticator.getClass());
@@ -173,7 +175,7 @@ class AuthenticatorFactoryImplTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     OboAuthenticator oboAuthenticator = factory.getOboAuthenticator();
     assertNotNull(oboAuthenticator);
     assertEquals(OboAuthenticatorCertImpl.class, oboAuthenticator.getClass());
@@ -188,7 +190,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator botAuth = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -200,7 +202,7 @@ class AuthenticatorFactoryImplTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator extAppAuthenticator = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorCertImpl.class, extAppAuthenticator.getClass());
@@ -215,7 +217,7 @@ class AuthenticatorFactoryImplTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
     config.getApp().getCertificate().setContent("certificate-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -229,7 +231,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
     config.getApp().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getApp().getPrivateKey().setContent("privatekey-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -245,7 +247,7 @@ class AuthenticatorFactoryImplTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -255,7 +257,7 @@ class AuthenticatorFactoryImplTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -265,7 +267,7 @@ class AuthenticatorFactoryImplTest {
 
     final BdkConfig config = new BdkConfig();
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -283,7 +285,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -299,7 +301,7 @@ class AuthenticatorFactoryImplTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -317,7 +319,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().setPrivateKey(privateKey);
     config.getApp().setPrivateKey(privateKey);
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -331,7 +333,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().setCertificatePath("/path/to/cert/cert.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -343,7 +345,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().getCertificate().setContent("certificate-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -355,7 +357,7 @@ class AuthenticatorFactoryImplTest {
     config.getBot().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
